mount-nv: support other filesystem types
Now mount-nv is able to automatically detect the type of filesystem on
the partition, to load block2dev only for JFFS2, and to properly use
and even format ext{2,3,4}. The detection code comes from the mtdpart
detection so it can still report some unexpected image types if any is
met, but that can also save from destroying a partition.

When called with --auto-format it will only format the partition after
a mount failure. When called with --format, it will always format it.
The default filesystem type is the same as already in place, or jffs2
for formating, or can be forced using -t.

Parts of this code could possibly be ported to native MTD devices by
removing the block2mtd stuff.