mount-nv: properly detect jffs2 when the superblock starts after empty blocks
On some large (?) file JFFS2 systems, it may happen that the superblock
starts anywhere. Eg it was observed at 31.5 MB on a 64 MB partition. This
could not be mounted because the resulting type was "empty". Now instead
if the type appears to be "empty", we skip all the first 0xFF and report
what follows. It's not rocket science but works well enough on a really
empty image.

If that's not enough, a future option could consist in returning "jffs2"
when "empty" was found but it would be better to avoid this if possible.